# docs/API.openapi.yaml
openapi: 3.0.0
info:
  title: BuildBrainOS API
  description: |
    Complete REST API specification for BuildBrainOS platform.
    All endpoints require JWT authentication in Authorization header.
    
    Base URLs:
    - Development: http://localhost:4000/api/v1
    - Production: https://api.buildbrainos.com/api/v1
    
    Authentication:
    - JWT Bearer Token in Authorization header
    - Refresh tokens valid for 30 days
    - Rate limiting: 1000 requests/hour per user
  version: 1.0.0
  contact:
    name: BuildBrainOS Support
    email: support@buildbrainos.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.buildbrainos.com/api/v1
    description: Production server
...
paths:
...
  /marketplace/bids:
    post:
      tags:
        - Marketplace
      summary: Submit project bid
      operationId: submitBid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - amount
                - timeline
              properties:
                projectId:
                  type: string
                amount:
                  type: number
                  format: double
                timeline:
                  type: object
                  properties:
                    startDate:
                      type: string
                      format: date
                    completionDate:
                      type: string
                      format: date
                proposal:
                  type: string
                  description: Detailed bid proposal text
                attachment:
                  type: string
                  format: binary
      responses:
        '201':
          description: Bid submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'
...
components:
  schemas:
...
    Bid:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        bidder:
          type: object
          properties:
            id:
              type: string
            companyName:
              type: string
            rating:
              type: number
        amount:
          type: number
          format: double
        status:
          type: string
          enum: [SUBMITTED, ACCEPTED, REJECTED, AWARDED]
        timeline:
          type: object
          properties:
            startDate:
              type: string
              format: date
            completionDate:
              type: string
              format: date
        proposal:
          type: string
        aiScore:
          type: number
          description: AI-generated bid quality score 0-100
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
...